chatmodels/requirements.txt
streamlit>=1.31.0
langchain-core>=0.1.0
langchain-huggingface>=0.0.1
python-dotenv>=1.0.0
